pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "delg"
version = "0.1.0"
description = "Dockerized DELG feature extractor with FastAPI and Python client."
authors = [
    { name = "Duje Giljanović", email = "giljanovic.duje@gmail.com" }
]
readme = "README.md"
license = { file = "LICENSE" }
dependencies = [
    "requests",
]

[tool.setuptools]
packages = ["delg"]

[tool.setuptools.package-data]
# include configs and model weights
"delg" = [
    "config_*.pbtxt",
    "parameters/*",
    "parameters/variables/*",
]

[project.optional-dependencies]
server = [
    "fastapi",
    "uvicorn",
    "tensorflow-cpu",
    "pillow",
    "numpy",
    "protobuf"
]
